Move `Value` to its own module
